requirements.txt
nest-asyncio
llama-index
llama-index-embeddings-gemini
llama-index-llms-gemini
llama-index-vector-stores-milvus
pymilvus
python-dotenv
tqdm>=4.66.4
pypdf>=3.17.4
fastapi
uvicorn[standard]
